Exception Handler

+++ src/main/java/nelsonssoares/ecomproductsapi/domain/dtos/SubCategoriaDTO.java
@@ -1,6 +1,6 @@
 package nelsonssoares.ecomproductsapi.domain.dtos;
 
 
-public record SubCategoriaDTO(String nomeSubCategoria, String descricao, Integer categoriaId) {
+public record SubCategoriaDTO(String nomeSubCategoria, String descricao, Long categoriaId) {
 
 }
